// DH.WebUI/src/entities/space-management/models/table-reservation-history.model.ts
import { ReservationStatus } from '../../../shared/enums/reservation-status.enum';

export interface ITableReservationHistory {
  id: number;
  userId: string;
  username: string;
  numberOfGuests: number;
  reservationDate: Date;
  createdDate: Date;
  isActive: boolean;
  status: ReservationStatus;
  isReservationSuccessful: boolean;
}